Change policy of applying rules.
Previously the policy of applying rules was that we emit the audit log
only when all applying rules return true regardless kind of rule. But in
this policy, we can never emit the DDL log or ROLE log when any object_id
rule is specified. So this changes that rule so that we apply the object_id
and object_type rules only when we're about to log regarding read, write
and misc.

Masahiko Sawada