Support timestamp rule.
The timestamp rule specifies the timestamp ranges that allows pgaudit
to emit the audit log if audit logging timestamp is in that range.
Similar to other rule the user can specify multiple timestamp ranges
and its equality. If the user specifies 'match' then pgaudit emits the
audit log when current timestamp is in any of timestamp ranges specified.
On the other hand, if the user specifies 'not match' then pgaudit emits
the audit log when current timestamp is not in all of the timestamp range
specified.

Because regression can be executed anytime, it's hard to test properly for
timestamp rule with current regression test frame work. So this dramatically
changes regression test frame work of pgaudit. Before invoking regression
test the pgaudit prepares audit configuration file and creating
conf/audit.conf from conf/audit.conf.org while setting appropriate timestamp
range for regression test at that time to config file and sql file,

Note that we expect that the timestamp rule regression test is excuted
firstly and it doesn't take long time until starting to regression test.