Introduce fixed format audit log message.
In current our design, the fixed format audit log is enough for
the user. But it could be changed in the future. In order to success
regression test regardless environment and executiion timing whatever,
we also introduced new parameter 'log_for_test' to output section.
If log_for_test is set, pgaudit don't write any information
depending on enviorment such as timestamp, virtual transaction id.
Because this parameter is completely for regression test, we're not
going to document this.

Masahiko Sawada