Support event trigger for regression test.
Previously, we ran regression test without event trigger functionality
of pgaudit. But since commit 126fa6d enable pgaudit to use pgaudit event
trigger we support it. To support it we somewhat change regression test
infrastructure dramatically so that it starts with 'empty.conf', includes
no rule, in order to install the pgaudit extension on all databases, and
changes the pgaudit config file to the properly file and then restart it.
All of these operations are done by init.sql script and we have to create
new database for regression test in init.sql when we want to add new
database.

Masahiko Sawada