Support audit_role filtering rule.
Using audit_role rule, the audit log is emitted only when current
user is listed in audit_role parameter if equality is specified.

It is important to ensure that pgaudit emits audit log even if a user
listed in audit_role, that is should be logged, do SET ROLE to change
to another role. It prevent spoofing by superuser.

Note that the behaviour of audit_role could be changed in future
commit so that it specifies the role group rather than one role name.
If the current user is a member of given role, emits audit log.
This change should satisfies issue #23.

Masahiko Sawada