Forcily change context of logging three GUC parameters to postmaster.
When starting, pgaudit checks if log_connections, log_disconnections
and log_replication_commands are enabled, and it doesn't start when
any of them is disabled. Therefore way we can guarantee that these
parameters are enabled when starting server. But it's possible to
change these during server running by do ALTER SYSTEM SET and
reloading configuration file. To prvent it, we forcibly change the
context of these to PGC_POSTMASTER in order to make these require
restarting server to change.

This changes fixes issue #31, reporeted by harada-toshi.